delete endpoint and delete service tests

+++ src/main/kotlin/com/abishake/fitnesstracker/repositories/EntryRepository.kt
@@ -3,7 +3,7 @@ package com.abishake.fitnesstracker.repositories
 import com.abishake.fitnesstracker.models.Entry
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface EntryRepository : JpaRepository<Entry, Int> {
+interface EntryRepository : JpaRepository<Entry, Long> {
     fun findByWorkoutId(workoutId: Int): List<Entry>
     fun findByExerciseId(exerciseId: Int): List<Entry>
 }
