+++ src/main/kotlin/com/abishake/fitnesstracker/repositories/ExerciseRepository.kt
@@ -4,6 +4,6 @@ import com.abishake.fitnesstracker.models.Exercise
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface ExerciseRepository : JpaRepository<Exercise, Int> {
+interface ExerciseRepository : JpaRepository<Exercise, Long> {
     fun findByName(name: String): Optional<Exercise>
 }
