+++ src/main/kotlin/com/abishake/fitnesstracker/repositories/WorkoutRepository.kt
@@ -4,10 +4,9 @@ import com.abishake.fitnesstracker.models.Workout
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import java.time.LocalDate
-import java.time.LocalDateTime
 
 @Repository
-interface WorkoutRepository : JpaRepository<Workout, Int> {
+interface WorkoutRepository : JpaRepository<Workout, Long> {
     fun findByName(name: String): List<Workout>
     fun findByCreatedAt(date: LocalDate): List<Workout>
 }
